+++ jxls/src/main/java/org/jxls/expression/ExpressionEvaluatorNoThreadLocalFactoryJexlImpl.java
@@ -7,6 +7,7 @@ public class ExpressionEvaluatorNoThreadLocalFactoryJexlImpl implements Expressi
 	private final boolean silent;
 	private final boolean strict;
 	private final JxlsJexlPermissions permissions;
+    private JexlContextFactory jexlContextFactory;
 
 	public ExpressionEvaluatorNoThreadLocalFactoryJexlImpl() {
 		this(false);
@@ -22,9 +23,28 @@ public class ExpressionEvaluatorNoThreadLocalFactoryJexlImpl implements Expressi
 		this.permissions = permissions;
 	}
 
+	public ExpressionEvaluatorNoThreadLocalFactoryJexlImpl(boolean silent, boolean strict, JxlsJexlPermissions permissions, JexlContextFactory jexlContextFactory) {
+		this.silent = silent;
+		this.strict = strict;
+		this.permissions = permissions;
+		this.jexlContextFactory = jexlContextFactory;
+	}
+
 	@Override
     public ExpressionEvaluator createExpressionEvaluator(final String expression) {
-        return expression == null ? new JexlExpressionEvaluatorNoThreadLocal(silent, strict, permissions)
-        		                  : new JexlExpressionEvaluatorNoThreadLocal(silent, strict, permissions, expression);
+		JexlExpressionEvaluatorNoThreadLocal ee = expression == null ? new JexlExpressionEvaluatorNoThreadLocal(silent, strict, permissions)
+				: new JexlExpressionEvaluatorNoThreadLocal(silent, strict, permissions, expression);
+        if (jexlContextFactory != null) {
+            ee.setJexlContextFactory(jexlContextFactory);
+        }
+		return ee;
+    }
+
+    public JexlContextFactory getJexlContextFactory() {
+        return jexlContextFactory;
+    }
+
+    public void setJexlContextFactory(JexlContextFactory jexlContextFactory) {
+        this.jexlContextFactory = jexlContextFactory;
     }
 }